BARK-37: Refactor showing of admin details into a AdminAccountView component, so we can reuse it for the delete screen.
