+++ src/app/admin/(logged-in)/tools/admins/view/[adminId]/page.tsx
@@ -3,9 +3,8 @@ import { getAuthenticatedAdminActor } from "@/lib/auth";
 import { RoutePath } from "@/lib/route-path";
 import { redirect } from "next/navigation";
 import { SimpleErrorPage } from "@/app/_components/simple-error-page";
-import Link from "next/link";
-import { Button } from "@/components/ui/button";
 import { BarkButton } from "@/components/bark/bark-button";
+import { AdminAccountView } from "../../_components/admin-account-view";
 
 export default async function Page(props: { params: { adminId: string } }) {
   const actor = await getAuthenticatedAdminActor();
@@ -20,44 +19,13 @@ export default async function Page(props: { params: { adminId: string } }) {
   if (error !== undefined) {
     return <SimpleErrorPage error={error} />;
   }
-  const {
-    adminName,
-    adminEmail,
-    adminPhoneNumber,
-    adminCanManageAdminAccounts,
-    adminCanManageVetAccounts,
-    adminCanManageUserAccounts,
-    adminCanManageDonors,
-  } = account;
-
   return (
     <div className="m-3 flex flex-col gap-6">
       <BarkBackLink
         href={RoutePath.ADMIN_TOOLS_ADMINS_LIST}
         text="Back to list"
       />
-      <div className="prose">
-        <h1>{adminName}</h1>
-        <p>Account ID: {adminId}</p>
-        <p>Email: {adminEmail}</p>
-        <p>Phone: {adminPhoneNumber}</p>
-        <h2>Permissions</h2>
-        <ul>
-          <_Permission
-            label="manage admin accounts"
-            value={adminCanManageAdminAccounts}
-          />
-          <_Permission
-            label="manage vet accounts"
-            value={adminCanManageVetAccounts}
-          />
-          <_Permission
-            label="manage user accounts"
-            value={adminCanManageUserAccounts}
-          />
-          <_Permission label="manage donors" value={adminCanManageDonors} />
-        </ul>
-      </div>
+      <AdminAccountView account={account} />
       <div className="flex flex-col gap-3 md:flex-row">
         <BarkButton
           className="w-full md:w-40"
@@ -76,16 +44,4 @@ export default async function Page(props: { params: { adminId: string } }) {
       </div>
     </div>
   );
-}
-
-function _Permission(props: { label: string; value: boolean }) {
-  const { label, value } = props;
-  if (value) {
-    return (
-      <li>
-        <b>Can {label}</b>
-      </li>
-    );
-  }
-  return <li>Cannot {label}</li>;
 }
